// apps/web/src/lib/api/generated/model/checkoutRequestPaymentMethod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews
 * OpenAPI spec version: 1.0.0
 */

export type CheckoutRequestPaymentMethod =
  (typeof CheckoutRequestPaymentMethod)[keyof typeof CheckoutRequestPaymentMethod];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CheckoutRequestPaymentMethod = {
  CASH: 'CASH',
  CARD: 'CARD',
  TRANSFER: 'TRANSFER',
} as const;