// apps/web/src/lib/api/generated/model/listInvoicesStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits
 * OpenAPI spec version: 1.0.0
 */

export type ListInvoicesStatus =
  (typeof ListInvoicesStatus)[keyof typeof ListInvoicesStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ListInvoicesStatus = {
  OPEN: 'OPEN',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED',
} as const;